Conflicts:
	lib/split/configuration.rb
	lib/split/helper.rb
	spec/configuration_spec.rb

+++ lib/split/helper.rb
@@ -3,33 +3,14 @@ module Split
     attr_accessor :ab_user
 
     def ab_test(experiment_name, control, *alternatives)
-      puts 'WARNING: You should always pass the control alternative through as the second argument with any other alternatives as the third because the order of the hash is not preserved in ruby 1.8' if RUBY_VERSION.match(/1\.8/) && alternatives.length.zero?
-      begin
-        experiment = Split::Experiment.find_or_create(experiment_name, *([control] + alternatives))
-        if experiment.winner
-          ret = experiment.winner.name
-        else
-          if forced_alternative = override(experiment.name, experiment.alternative_names)
-            ret = forced_alternative
-          else
-            clean_old_versions(experiment)
-            begin_experiment(experiment) if exclude_visitor? or not_allowed_to_test?(experiment.key)
 
-            if ab_user.get_key(experiment.key) 
-              ret = ab_user.get_key(experiment.key)
+      puts 'WARNING: You should always pass the control alternative through as the second argument with any other alternatives as the third because the order of the hash is not preserved in ruby 1.8' if RUBY_VERSION.match(/1\.8/) && alternatives.length.zero?
+      ret = if Split.configuration.enabled
+              experiment_variable(alternatives, control, experiment_name)
             else
-              alternative = experiment.next_alternative
-              alternative.increment_participation
-              begin_experiment(experiment, alternative.name)
-              ret = alternative.name
+              control_variable(control)
             end
-          end
-        end
-      rescue Errno::ECONNREFUSED => e
-        raise unless Split.configuration.db_failover
-        Split.configuration.db_failover_on_db_error.call(e)
-        ret = Hash === control ? control.keys.first : control
-      end
+
       if block_given?
         if defined?(capture) # a block in a rails view
           block = Proc.new { yield(ret) }
@@ -44,9 +25,9 @@ module Split
     end
 
     def finished(experiment_name, options = {:reset => true})
-      return if exclude_visitor?
+      return if exclude_visitor? or !Split.configuration.enabled
       return unless (experiment = Split::Experiment.find(experiment_name))
-      if alternative_name = ab_user.get_key(experiment.key)
+      if alternative_name = ab_user[experiment.key]
         alternative = Split::Alternative.new(alternative_name, experiment_name)
         alternative.increment_completion
         ab_user.delete_key(experiment_name) if options[:reset]
@@ -57,11 +38,7 @@ module Split
     end
 
     def override(experiment_name, alternatives)
-      begin
-        params[experiment_name] if alternatives.include?(params[experiment_name])
-      rescue
-        nil
-      end
+      params[experiment_name] if defined?(params) && alternatives.include?(params[experiment_name])
     end
 
     def begin_experiment(experiment, alternative_name = nil)
@@ -82,7 +59,7 @@ module Split
     end
 
     def doing_other_tests?(experiment_key)
-      ab_user.get_keys.reject{|k| k == experiment_key}.length > 0
+      ab_user.get_keys.reject { |k| k == experiment_key }.length > 0
     end
 
     def clean_old_versions(experiment)
@@ -93,7 +70,7 @@ module Split
 
     def old_versions(experiment)
       if experiment.version > 0
-        ab_user.get_keys.select{|k| k.match(Regexp.new(experiment.name))}.reject{|k| k == experiment.key}
+        ab_user.get_keys.select { |k| k.match(Regexp.new(experiment.name)) }.reject { |k| k == experiment.key }
       else
         []
       end
@@ -114,5 +91,44 @@ module Split
         false
       end
     end
+
+
+    protected
+
+    def control_variable(control)
+      Hash === control ? control.keys.first : control
+    end
+
+    def experiment_variable(alternatives, control, experiment_name)
+      begin
+        experiment = Split::Experiment.find_or_create(experiment_name, *([control] + alternatives))
+        if experiment.winner
+          ret = experiment.winner.name
+        else
+          if forced_alternative = override(experiment.name, experiment.alternative_names)
+            ret = forced_alternative
+          else
+            clean_old_versions(experiment)
+            begin_experiment(experiment) if exclude_visitor? or not_allowed_to_test?(experiment.key)
+
+            if ab_user[experiment.key]
+              ret = ab_user[experiment.key]
+            else
+              alternative = experiment.next_alternative
+              alternative.increment_participation
+              begin_experiment(experiment, alternative.name)
+              ret = alternative.name
+            end
+          end
+        end
+      rescue Errno::ECONNREFUSED => e
+        raise unless Split.configuration.db_failover
+        Split.configuration.db_failover_on_db_error.call(e)
+        ret = control_variable(control)
+      end
+      ret
+    end
+
   end
+
 end
